Assistant checkpoint: Fix vineyard table reloading issue
Assistant generated file changes:
- js/overlays/mainpages/vineyardoverlay.js: Clear table container before adding new table

---

User prompt:

Okay that works now. However there seems to be a bug in the vineyard table.

at first load of vineyardoverlay the vineyardtable loads fine and is populated with the farmland class. However, if vineyardoverlay is reloaded (IE link is clicked again) we get a unpopulated table. Any suggestion to why this happens ?