Assistant checkpoint: Add Must container selection to crushing process
Assistant generated file changes:
- js/overlays/crushingOverlay.js: Add Must container selection to crushing overlay

---

User prompt:

Okay i see the problem. When we crushed the Grapes we changed the Status to must, but we did not change the Storage.

In the Crushingoverlay. We need to let the user select which container the crushed must should go into. And we will allow only tools that has supportedResources='Must'

This should work much like the selection process of canHarvest(). Where we find the storage appropiate, check that there is empty storage and then give the user the possible options to select from.