Assistant checkpoint: Fix inventory management during harvest
Assistant generated file changes:
- js/resource.js: Fix storage validation in addResource
- js/overlays/harvestOverlay.js: Add vintage check in harvest function

---

User prompt:

There seems to be a issue with the way manipulate the inventory when harvesting  It seems that we create a new resource if there is a new vintage, but we also seem to manipulate the State of resources allready in the inventory. Like changing the state of a earlier vintage. IE when harvesting the 2024 vintage, we create a new resource barbera, 2024, Grape. However we also maniputale the exxcisting 2023 vintage, that has allready been transformet into Must, back to Grapes, and stack both vintage in the same container.