Assistant checkpoint: Fix bottle info to include field name and other wine data
Assistant generated file changes:
- js/resource.js: Fix getBottleInfo to include all necessary wine data

---

User prompt:

No. We absolutly dont want to just use a random value. We want to fix why there is no farmland. And that is because we are not using bottleswine.fieldname. We might be using ressource.fieldname or this.fieldname. You should look in resource.js or in wineprocesing.js to find out