Assistant checkpoint: Update harvest overlay to show large numbers in tons
Assistant generated file changes:
- js/overlays/harvestOverlay.js: Update number formatting to use tons for large values

---

User prompt:

There is something about the formatNumbers that is not working correctly. It should list big numbers in t not in kg. But it does not seem to. This should allready be implemented as it works for vineyard overlay under expected yields

We do not want to change the formatNumber function as it serve other purposes as well rather we want to do the same way as we do in vineyardoverlay

<td>${farmlandYield(farmland) >= 1000 ? formatNumber(farmlandYield(farmland)/1000, 2) + ' t' : formatNumber(farmlandYield(farmland), 2) + ' kg'}</td>