Assistant checkpoint: Update mainview overlay to use CSS variables for widths
Assistant generated file changes:
- css/overlay.css: Make mainview width responsive to sidebar states

---

User prompt:

Not yet. In the right we also have a panel that takes up 250px. Leaving the rest of the page (-500px) for the mainview. If we were also to introduce a collaps/expand on the panel. We would then need a flexsible area for the mainview. I think this is controlled by overlay.css as of now?