Assistant checkpoint: Add harvest amount check to prevent zero quantity entries
Assistant generated file changes:
- js/overlays/harvestOverlay.js: Add check for positive harvest amount before creating inventory entry

---

User prompt:

There seems to be a bug in the harvest, or performharvest function. Where we seems to be creating a entry to the storage/inventory for the harvested grape, even if 0 grapes was harvest (IE 0 workapplied = no worker assignet or worker with 0 skill assignet). This eventually leads to a entry with 0 storaged quantity with NaN quality. Breaking the system

We want to make sure we do not put anything in the inventiory until there is actually harvested something. IE >0 grape harvested