Assistant checkpoint: Fix task serialization for crushing task persistence
Assistant generated file changes:
- js/database/adminFunctions.js: Fix task serialization for crushing tasks
Fix task serialization for crushing tasks

---

User prompt:

we get this error, when we reload the page, while there are a active crushing task. This only accurse when the page is reloaded. Suggesting that its somekind of storeing error. Are we sure we store in the right way. Maybe related to how we handle data in adminFunctions, or ?

crushingOverlay.js:406 Uncaught TypeError: Cannot read properties of undefined (reading 'resource')
    at performCrushing (crushingOverlay.js:406:48)
    at Object.callback (adminFunctions.js:485:9)
    at taskManager.js:117:22
    at Map.forEach (<anonymous>)
    at TaskManager.processWeek (taskManager.js:84:20)
    at incrementWeek (endDay.js:21:17)
    at HTMLButtonElement.<anonymous> (mainofficeoverlay.js:37:13)