Assistant checkpoint: Simplified task storage by removing special case for crushing
Assistant generated file changes:
- js/database/adminFunctions.js: Remove special case for crushing tasks

---

User prompt:

Okay i dont want to have a special case for crushing. If we cant store it as a params, then we will need to create a new entry called selected grape. Im not sure why we need this, why cant we just use the ressource.name as we do in a lot of other data handling ?