Assistant checkpoint: Refactor crushing functions to use inventory instance
Assistant generated file changes:
- js/overlays/crushingOverlay.js: Refactor crushing function to use inventory instance, Update performCrushing to use inventory instance

---

User prompt:

In crushing() and in performcrushing (). We handle data by creating the selectedGrape.dataset. This is causing problems, we dont do this in fermentation() and performfermentation() instead we seems to use the data from the inventoryinstance.

I want to change crushing and performcrushing to be more like fermentation and performfermentation