Assistant checkpoint: Simplify crushing task storage logic
Assistant generated file changes:
- js/overlays/crushingOverlay.js: Simplify crushing task creation, Simplify performCrushing function

---

User prompt:

okay we now get Grapes not found in storage, when we reload the page while there is a active crushing task. This is because we are trying to acces a object stored in a JSON stringify. I DONT want to try and make the object stored better. I want to eliminate the need to store values that we have in the object. I think this is because of the way we construct grapeResource. This is unnessesary complex compared to how we do in fermentation().