Assistant checkpoint: Implement date-triggered task system with Bookkeeping task
Assistant generated file changes:
- js/administration.js: Replace old bookkeeping task with new recurring system
- js/taskManager.js: Add date-triggered task handling

---

User prompt:

I want to create a new kind of task, that is activated by date (ingame date)  but uses the same taskManager system as the normal task, that is activated via button clicks. We can call these Running Task

We got some old code in adminstration.js. the bookkeepingTaskFunction. We will rename this to just Bookeeping() and use the general ideas about that function. That every 1. week (IE every time a new season starts) this task is activated. If the previus Bookkeeping task is not done by the new trigging the new work will be added to the previus and with a 10% panelty (Some form of this is allready in the old code)

The task total work will be calculated as the number of transaction in the previus season (IE 1. week of summer 2025, we will have the number of transaction in spring 2025 * 10, as the total work, for a task called Bookkeeping Spring 2025) + eventually not done work in older task *1,1 )

When the work is done, we will give a confirmation addconsolemsg to the user that this seasons bookkepping task is done. If the task is not completed by the 1. week of next season in addition to spilloverwork, we still give a penalty to prestige